+++ packages/openai/src/responses/openai-responses-api.ts
@@ -380,7 +380,12 @@ export const openaiResponsesChunkSchema = lazyValidator(() =>
                 type: z.literal('search'),
                 query: z.string().nullish(),
                 sources: z
-                  .array(z.object({ type: z.literal('url'), url: z.string() }))
+                  .array(
+                    z.discriminatedUnion('type', [
+                      z.object({ type: z.literal('url'), url: z.string() }),
+                      z.object({ type: z.literal('api'), name: z.string() }),
+                    ]),
+                  )
                   .nullish(),
               }),
               z.object({
@@ -596,7 +601,12 @@ export const openaiResponsesResponseSchema = lazyValidator(() =>
                 type: z.literal('search'),
                 query: z.string().nullish(),
                 sources: z
-                  .array(z.object({ type: z.literal('url'), url: z.string() }))
+                  .array(
+                    z.discriminatedUnion('type', [
+                      z.object({ type: z.literal('url'), url: z.string() }),
+                      z.object({ type: z.literal('api'), name: z.string() }),
+                    ]),
+                  )
                   .nullish(),
               }),
               z.object({