+++ packages/cohere/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/cohere
 
+## 2.0.18
+
+### Patch Changes
+
+- 51aa5de: backport: test server
+- Updated dependencies [51aa5de]
+  - @ai-sdk/provider-utils@3.0.16
+
+## 2.0.17
+
+### Patch Changes
+
+- Updated dependencies [f2da310]
+  - @ai-sdk/provider-utils@3.0.15
+
 ## 2.0.16
 
 ### Patch Changes