+++ packages/google/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/google
 
+## 2.0.28
+
+### Patch Changes
+
+- 9a999c4: fix(@ai-sdk/google): Make title field optional in grounding metadata schema
+
+## 2.0.27
+
+### Patch Changes
+
+- 51aa5de: backport: test server
+- Updated dependencies [51aa5de]
+  - @ai-sdk/provider-utils@3.0.16
+
+## 2.0.26
+
+### Patch Changes
+
+- Updated dependencies [f2da310]
+  - @ai-sdk/provider-utils@3.0.15
+
 ## 2.0.25
 
 ### Patch Changes