+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/provider-utils
 
+## 3.0.16
+
+### Patch Changes
+
+- 51aa5de: backport: test server
+
+## 3.0.15
+
+### Patch Changes
+
+- f2da310: chore(provider-utils): upgrade eventsource-parser to 3.0.6
+
 ## 3.0.14
 
 ### Patch Changes