+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/valibot
 
+## 1.0.16
+
+### Patch Changes
+
+- 51aa5de: backport: test server
+- Updated dependencies [51aa5de]
+  - @ai-sdk/provider-utils@3.0.16
+
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [f2da310]
+  - @ai-sdk/provider-utils@3.0.15
+
 ## 1.0.14
 
 ### Patch Changes